+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 * 1.5.2 - 2024/12/05
   * Add Klarna payment compatibility
   * Add log visibility in backoffice
+  * Fix billingDay for NX payment
 
 ---
 * 1.5.1 - 2024/10/24